pyproject.toml
[tool.poetry]
name = "learn-fastapi"
version = "0.1.0"
description = ""
authors = ["subratamondal1 <subratasubha2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
ipykernel = "^6.29.5"
httpie = "^3.2.3"
sqlalchemy = {extras = ["asyncio", "mypy"], version = "^2.0.36"}
pydantic = "^2.9.2"
pymongo-stubs = "^0.2.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
